+++ CHANGELOG.md
@@ -50,6 +50,8 @@ such as preferences, LocalStorage, and Dashboards.  This has been enabled on `Gr
 * The mobile `Select` component supports the `enableFilter` and `enableCreate` props.
 * `DashContainerModel` supports new `layoutLocked`, `contentLocked` and `renameLocked` modes.
 * `DimensionChooser` now has the ability to persist its value and history separately.
+* Enhance Hoist Admin's Activity Tracking tab.
+* Enhance Hoist Admin's Client Error tab.
 
 ### 💥 Breaking Changes